+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 - fix: CKV2_AWS_28 Ensure public facing ALB are protected by WAF
-- fix: CKV_AWS_111 #Ensure IAM policies does not allow write access without constraints
+- feat: more than one security group for ecs service
 - feat: Add EC2 usage example
 - feat: Possibly use lb module for load-balancer resource
 - feat: Review ecs-service arguments, add and test those missing.
@@ -14,9 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: Add more options for module cloudwatch log group
 - feat: Exclusively use acm certificate (not self_signed_cert) for complete example
 - feat: use updated vpc source in examples
-- feat: add supporting resources for testing
 - feat: use updated ecs cluster source for ecs cluster
 - feat: consolidate ecs cluster module and ecs service module into one
+
+## [1.1.5] - 2022-10-20
+### Changes
+- fix: CKV_AWS_111 #Ensure IAM policies does not allow write access without constraints
+- fix: CKV_AWS_158 #Ensure that CloudWatch Log Group is encrypted by KMS
+- hotfix: deployment controller in minimum example
+- fix: egress rules for fargate example
 
 ## [1.1.4] - 2022-10-10
 ### Changes
@@ -65,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: feature update.
 - feat: initial code commit
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-ecs-service/compare/1.1.4...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-ecs-service/compare/1.1.5...HEAD
+[1.1.5]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.1.5
 [1.1.4]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.1.4
 [1.1.3]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.1.3
 [1.1.2]: https://github.com/boldlink/terraform-aws-ecs-service/releases/tag/1.1.2